Dropping `node-objc`, not gonna work with how dynamic Objective-C is.
Instead I'm using `node-ffi` directly in order to dynamically generate the proper
`objc_msgSend` function as needed for the various types of selectors.

Also dropping `xml2js`. It sucks. Mainly because when there's only 1 node returned
with a given name, it returns an Object representing that. But if there's
multiple of the same name, then it's an Array of Objects! This may sound "nice"
but it makes it difficult to program for. Now using isaacs `sax-js` directly.