First attempt at caching the object wrap instance retrieval.
Part of #4.

This uses `objc_getAssociatedObject()` and `objc_setAssociatedObject()`.
The actualy Function wrap instance is stored in a Pointer with putObject(),
and retrieved in later times, for the same object to be returned (strict
equality and all that jazz).

However, things don't seem entirely stable with this patch. Needs more testing...
It's possible that before the key is initalized, there's a possibility of picking
up garbage data (i.e. doesn't always return NULL). And randomly it ends up not
being null and when getObject() is called, bad stuff happens. Like I said, needs
more testing...